Initial small commit

+++ folly/SpookyHashV1.cpp
@@ -196,7 +196,7 @@ void SpookyHashV1::Hash128(
     remainder = (length - ((const uint8_t *)end-(const uint8_t *)message));
     memcpy(buf, end, remainder);
     memset(((uint8_t *)buf)+remainder, 0, sc_blockSize-remainder);
-    ((uint8_t *)buf)[sc_blockSize-1] = remainder;
+    ((uint8_t *)buf)[sc_blockSize-1] = (uint8_t)remainder;
     Mix(buf, h0,h1,h2,h3,h4,h5,h6,h7,h8,h9,h10,h11);
 
     // do some final mixing
